Fix coverity issue in encoder_pkey
CID 1515953

Since sk_OPENSSL_CSTRING_num() can return -1 the return result needs to be
stored as a int - not a size_t.
In practice it would not hit this case.

+++ crypto/encode_decode/encoder_pkey.c
@@ -103,7 +103,7 @@ static void collect_encoder(OSSL_ENCODER *encoder, void *arg)
      */
     if ((data->keymgmt_prov == prov) == data->flag_find_same_provider) {
         void *provctx = OSSL_PROVIDER_get0_provider_ctx(prov);
-        size_t i, end_i = sk_OPENSSL_CSTRING_num(data->names);
+        int i, end_i = sk_OPENSSL_CSTRING_num(data->names);
         int match;
 
         for (i = 0; i < end_i; i++) {
